Split enrichment in to 2 distinct phases
The first phase relies only on clang.Cursor or clang.Type for
information. In same cases it adds a clanf.Type to a struct type.

The second phase uses the api type instance. The api was fully
populated during phase one, so it can be used to find records, enums,
and the like.