Only build skia when using script with -s flag
Avoids rebuilding lib/libskia.a et al when iterating on generation
code. When the library files are built they include a timestamp,
which means that they are not produced deterministically. That means
that they'll need to be commited, which is not desired when they
have not really changed (and would bloat the repo).