Only copy libraries following a 'clean' build
This is a better approach than a previous attempt (b22a164) that
tried to force timestamps. That approach didn't work, it was not in
fact deterministic.

Now, if the '-c' (or '-C') flag is used with the build script the
library copy directory (and header file directory) are removed. And if
the directory is not present when the build script runs the directory
is created and the files are copied. The means that committing new
library (or header) files is a deliberate act.